fix: 로그인 헤더 설정 오류 수정
